backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.3.2
pydantic==2.6.1
python-dotenv==1.0.1
